requirements.txt
requests
python-dotenv
asyncpg
psycopg
psycopg[binary] 
sqlalchemy
pandas
scikit-learn
matplotlib
tensorflow